requirements.txt
asgiref==3.5.0
Django==4.0.3
geopy==2.2.0
python-decouple==3.6
sqlparse==0.4.2
tzdata==2021.5
Pillow==9.0.1

